refactor insert

+++ Task3/db/DDL/tables/create_films_table.sql
@@ -4,7 +4,7 @@ create table if not exists films
 (
     film_id       int           not null,
     title         varchar(200)  null,
-    year          int(4)        null,
+    year          smallint(4)        null,
     genres        varchar(100)  null,
     ratings_count int default 0 null,
     ratings_sum   int default 0 null,
