+++ Task3/db/DDL/tables/create_filtered_table.sql
@@ -3,7 +3,7 @@ use films_catalog;
 create table if not exists filtered_films
 (
     title  varchar(200) null,
-    year   int(4)       null,
+    year   smallint(4)       null,
     genres varchar(100) null,
     rating float        null
 );
